test-requirements.txt
odoo-addon-base_queued_notify @ git+https://github.com/sygel-technology/sy-server-tools.git@refs/pull/7/head#subdirectory=base_queued_notify
